refactored the delete functionality to refresh the stale data when deleting the collection
